work if only server.jar and no properties file

+++ server/Service/Services/ServerPropertiesService.cs
@@ -2,7 +2,6 @@
 using Service.InputModels;
 using Service.IServices;
 using System.Reflection;
-using System.Text.Json.Serialization;
 using System.Text.RegularExpressions;
 
 namespace Service.Services;
@@ -13,35 +12,37 @@ public class ServerPropertiesService : IServerPropertiesService
     {
         var filePath = Path.Combine(Util.GetServerPath(serverName), "server.properties");
         // Read in the file 
-        string fileContents = File.ReadAllText(filePath);
+        if (File.Exists(filePath)){
+            string fileContents = File.ReadAllText(filePath);
 
 
-        Type classType = serverProperties.GetType();
-        foreach(PropertyInfo property in classType.GetProperties(BindingFlags.Public | BindingFlags.Instance))
-        {
-            var propertyValue = property.GetValue(serverProperties);
-            if(propertyValue != null)
+            Type classType = serverProperties.GetType();
+            foreach(PropertyInfo property in classType.GetProperties(BindingFlags.Public | BindingFlags.Instance))
             {
-                var propertyName = GetPropertyName(property);
-                // Use a regular expression to find the line with the property we want to change
-                string pattern = $"{propertyName}=(.*)";
-                Regex regex = new Regex(pattern);
-                Match match = regex.Match(fileContents);
+                var propertyValue = property.GetValue(serverProperties);
+                if(propertyValue != null)
+                {
+                    var propertyName = GetPropertyName(property);
+                    // Use a regular expression to find the line with the property we want to change
+                    string pattern = $"{propertyName}=(.*)";
+                    Regex regex = new Regex(pattern);
+                    Match match = regex.Match(fileContents);
 
-                if (match.Success)
-                {
-                    // Replace the old value with the new value
-                    string oldValue = match.Groups[1].Value;
-                    string newLine = $"{propertyName}={propertyValue}";
-                    fileContents = fileContents.Replace($"{propertyName}={oldValue}", newLine);
+                    if (match.Success)
+                    {
+                        // Replace the old value with the new value
+                        string oldValue = match.Groups[1].Value;
+                        string newLine = $"{propertyName}={propertyValue}";
+                        fileContents = fileContents.Replace($"{propertyName}={oldValue}", newLine);
+                    }
+                    else
+                    {
+                        throw new ArgumentException($"Property {property.Name} not found in file {filePath}");
+                    }
                 }
-                else
-                {
-                    throw new ArgumentException($"Property {property.Name} not found in file {filePath}");
-                }
-            }
-        } // Write the updated file contents back to the file
-        File.WriteAllText(filePath, fileContents);
+            } // Write the updated file contents back to the file
+            File.WriteAllText(filePath, fileContents);
+        }
     }
 
     private string GetPropertyName(PropertyInfo prop)
